requirements.txt
annotated-types==0.7.0
anyio==4.4.0
av==12.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.2.1
distro==1.9.0
faster-whisper==1.0.2
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.6.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.1
groq==0.6.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
openai==1.30.1
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.7.3
pydantic_core==2.18.4
pyparsing==3.1.2
pyperclip==1.8.2
python-dotenv==1.0.1
pyttsx3==2.90
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SpeechRecognition==3.10.1
sympy==1.12.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
